requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
opencv-python>=4.13.0
numpy>=2.4.3
psycopg2-binary>=2.9.11
minio==7.1.14
python-dotenv==1.0.0
Pillow>=11.3.0
requests==2.31.0
